requirements.txt
tensorflow>=1.15.0
matplotlib>=3.2.2
requests>=2.24.0

google.auth>=1.14.1
